+++ app/src/main/java/com/wanderwildwood/kirinuki/model/opml/OPMLImporter.kt
@@ -85,6 +85,7 @@ open class OPMLImporter(
                     linkOpenerFromString(value),
                 )
             UserSettings.SETTING_TEXT_SCALE -> settingsStore.setTextScale(value.toFloatOrNull() ?: 1.0f)
+            UserSettings.SETTING_OPEN_TITLE_IN_BROWSER -> settingsStore.setOpenTitleInBrowser(value.toBoolean())
             UserSettings.SETTING_IS_MARK_AS_READ_ON_SCROLL -> settingsStore.setIsMarkAsReadOnScroll(value.toBoolean())
             UserSettings.SETTING_READALOUD_USE_DETECT_LANGUAGE -> settingsStore.setUseDetectLanguage(value.toBoolean())
             UserSettings.SETTING_MAX_LINES -> settingsStore.setMaxLines((value.toIntOrNull() ?: 1).coerceAtLeast(1))